update error handling sysytem
