--- __main__.py
@@ -1,3 +1,0 @@
-from s3sync_util.main import cli
-
-cli()
