--- __version__.py
@@ -1,1 +1,0 @@
-__version__ = "2.3.1"
