+++ s3sync_util/commands/common.py
@@ -1,6 +1,6 @@
 import os
 import boto3
-
+from ..logger import logger
 from boto3.s3.transfer import TransferConfig
 
 config = TransferConfig(multipart_threshold=1024 * 25, 
@@ -24,8 +24,9 @@ def get_total_upload_objects(directory:str, exclude_list:list) -> int:
         for file in files:
             if file not in exclude_list:
                 total_objects += 1
-    return total_objects
 
+    return total_objects
+    
 def get_total_download_objects(bucket:str, prefix:str) -> int:
     """Count the total number of objects (files and directories) in an S3 bucket with a given prefix.
 
@@ -40,6 +41,7 @@ def get_total_download_objects(bucket:str, prefix:str) -> int:
 
     response = s3.list_objects_v2(Bucket=bucket, Prefix=prefix)
     total_objects = len(response.get('Contents', []))
+
     return total_objects
 
 def format_time(seconds:int) -> str:
@@ -53,4 +55,5 @@ def format_time(seconds:int) -> str:
     """
     m, s = divmod(seconds, 60)
     h, m = divmod(m, 60)
+    logger.info("calculated download and upload time.")
     return f"{int(h):02d}:{int(m):02d}:{int(s):02d}"
